Move hooks to base test
